requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
lxml
playwright
pandas
schedule
supabase
python-dotenv
